requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
requests
langchain-google-genai
google-generativeai
langchain-cohere
sentence-transformers
langchain-huggingface
langchain-chroma
chromadb
faiss-cpu
langchain-qdrant
qrant_client
streamlit
langchain-groq
fastapi
uvicorn
langserve
arxiv
wikipedia
langchainhub
duckduckgo-search
psycopg2
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain_nvidia_ai_endpoints
crewai
crewai-tools
yt_dlp==2023.11.14
youtube-transcript-api==0.6.1